refactor backend to better handle different schema versions
