+++ projects/app/test/server/lib/replicache.test.ts
@@ -1,978 +1,48 @@
-import {
-  srsStateFromFsrsState,
-  supportedSchemas,
-  v7,
-} from "#data/rizzleSchema.ts";
-import { glossToHanziWord } from "#data/skills.ts";
-import { xmin } from "#server/lib/db.ts";
-import {
-  computeCvrEntities,
-  fetchMutations,
-  pull,
-  push,
-} from "#server/lib/replicache.ts";
-import * as s from "#server/schema.ts";
-import { nextReview, Rating } from "#util/fsrs.ts";
-import { nanoid } from "#util/nanoid.ts";
-import { invariant } from "@haohaohow/lib/invariant";
-import { eq } from "drizzle-orm";
+import { pull, push } from "#server/lib/replicache.ts";
 import assert from "node:assert/strict";
 import { test } from "node:test";
-import { createUser, withDbTest, withTxTest } from "./dbHelpers";
+import { withTxTest } from "./dbHelpers";
 
 await test(`${push.name} suite`, async (t) => {
-  for (const schema of supportedSchemas) {
-    await t.test(`schema ${schema.version}`, async (t) => {
-      const txTest = withTxTest(t, { isolationLevel: `repeatable read` });
+  const txTest = withTxTest(t, { isolationLevel: `repeatable read` });
 
-      await test(`database transaction isolation level`, async (t) => {
-        const txTest = withTxTest(t, { isolationLevel: `read committed` });
-
-        await txTest(`fails when using the default`, async (tx) => {
-          const result = push(tx, `1`, {
-            profileId: ``,
-            clientGroupId: ``,
-            pushVersion: 1,
-            schemaVersion: schema.version,
-            mutations: [],
-          });
-
-          await assert.rejects(result, /transaction_isolation/);
-        });
+  await txTest(
+    `returns correct error for invalid schema version`,
+    async (tx) => {
+      const result = await push(tx, `1`, {
+        profileId: ``,
+        clientGroupId: ``,
+        pushVersion: 1,
+        schemaVersion: `666`,
+        mutations: [],
       });
 
-      await txTest(`handles no mutations`, async (tx) => {
-        await push(tx, `1`, {
-          profileId: ``,
-          clientGroupId: ``,
-          pushVersion: 1,
-          schemaVersion: schema.version,
-          mutations: [],
-        });
+      assert.deepEqual(result, {
+        error: `VersionNotSupported`,
+        versionType: `schema`,
       });
-
-      await txTest(
-        `only allows a client group if it matches the user`,
-        async (tx) => {
-          const user1 = await createUser(tx);
-          const user2 = await createUser(tx);
-
-          // Create a client group
-          const [clientGroup] = await tx
-            .insert(s.replicacheClientGroup)
-            .values([
-              {
-                id: `1`,
-                userId: user1.id,
-                cvrVersion: 1,
-                schemaVersion: schema.version,
-              },
-            ])
-            .returning();
-          invariant(clientGroup != null);
-
-          const mut = {
-            id: 1,
-            name: `noop`,
-            args: {},
-            timestamp: 10_123,
-            clientId: `c0f86dc7-4d49-4f37-a25b-4d06c9f1cb37`,
-          };
-
-          // User 2 doesn't own the clientGroup
-          await assert.rejects(
-            push(tx, user2.id, {
-              profileId: ``,
-              clientGroupId: clientGroup.id,
-              pushVersion: 1,
-              schemaVersion: schema.version,
-              mutations: [mut],
-            }),
-          );
-
-          // User 1 does own the clientGroup
-          await push(tx, user1.id, {
-            profileId: ``,
-            clientGroupId: clientGroup.id,
-            pushVersion: 1,
-            schemaVersion: schema.version,
-            mutations: [mut],
-          });
-        },
-      );
-
-      await txTest(
-        `creates a client group and client if one doesn't exist`,
-        async (tx) => {
-          const clientId = `clientid`;
-          const clientGroupId = `clientgroupid`;
-
-          const user = await createUser(tx);
-
-          const mut = {
-            id: 1,
-            name: `noop`,
-            args: {},
-            timestamp: 1,
-            clientId,
-          };
-
-          await push(tx, user.id, {
-            profileId: ``,
-            clientGroupId,
-            pushVersion: 1,
-            schemaVersion: schema.version,
-            mutations: [mut],
-          });
-
-          const client = await tx.query.replicacheClient.findFirst({
-            where: (t, { eq }) => eq(t.id, clientId),
-          });
-          assert.equal(client?.id, clientId);
-
-          const clientGroup = await tx.query.replicacheClientGroup.findFirst({
-            where: (t, { eq }) => eq(t.id, clientGroupId),
-          });
-          assert.equal(clientGroup?.id, clientGroupId);
-        },
-      );
-
-      await txTest(`skips already processed mutations`, async (tx) => {
-        const user = await createUser(tx);
-
-        // Create a client group
-        const [clientGroup] = await tx
-          .insert(s.replicacheClientGroup)
-          .values([{ userId: user.id, schemaVersion: schema.version }])
-          .returning();
-        invariant(clientGroup != null);
-
-        // Create a client
-        const [client] = await tx
-          .insert(s.replicacheClient)
-          .values([{ clientGroupId: clientGroup.id, lastMutationId: 1 }])
-          .returning();
-        invariant(client != null);
-
-        const mut = {
-          id: client.lastMutationId, // use the same ID
-          name: schema.rateSkill._def.alias!,
-          args: schema.rateSkill.marshalArgs({
-            id: nanoid(),
-            skill: glossToHanziWord(`我:i`),
-            rating: Rating.Good,
-            durationMs: null,
-            now: new Date(),
-          }),
-          timestamp: 1,
-          clientId: client.id,
-        };
-
-        await push(tx, user.id, {
-          profileId: ``,
-          clientGroupId: clientGroup.id,
-          pushVersion: 1,
-          schemaVersion: schema.version,
-          mutations: [mut],
-        });
-
-        assert.deepEqual(
-          await tx.query.skillState.findMany({
-            where: (t, { eq }) => eq(t.userId, user.id),
-          }),
-          // The mutation SHOULDN'T have done anything, it should be skipped.
-          [],
-        );
-      });
-
-      await txTest(`does not process mutations from the future`, async (tx) => {
-        const user = await createUser(tx);
-
-        // Create a client group
-        const [clientGroup] = await tx
-          .insert(s.replicacheClientGroup)
-          .values([{ userId: user.id, schemaVersion: schema.version }])
-          .returning();
-        invariant(clientGroup != null);
-
-        // Create a client
-        const [client] = await tx
-          .insert(s.replicacheClient)
-          .values([{ clientGroupId: clientGroup.id, lastMutationId: 1 }])
-          .returning();
-        invariant(client != null);
-
-        const mut = {
-          id: client.lastMutationId + 2,
-          name: ``,
-          args: {},
-          timestamp: 1,
-          clientId: client.id,
-        };
-
-        await assert.rejects(
-          push(tx, user.id, {
-            profileId: ``,
-            clientGroupId: clientGroup.id,
-            pushVersion: 1,
-            schemaVersion: schema.version,
-            mutations: [mut],
-          }),
-        );
-      });
-
-      await txTest(
-        `invalid mutations must still update client.lastMutationID`,
-        async (tx) => {
-          const user = await createUser(tx);
-
-          // Create a client group
-          const [clientGroup] = await tx
-            .insert(s.replicacheClientGroup)
-            .values([{ userId: user.id, schemaVersion: schema.version }])
-            .returning();
-          invariant(clientGroup != null);
-
-          // Create a client
-          const [client] = await tx
-            .insert(s.replicacheClient)
-            .values([{ clientGroupId: clientGroup.id, lastMutationId: 1 }])
-            .returning();
-          invariant(client != null);
-
-          const mut = {
-            id: client.lastMutationId + 1,
-            name: `invalidMutation`,
-            args: {},
-            timestamp: 1,
-            clientId: client.id,
-          };
-
-          await push(tx, user.id, {
-            profileId: ``,
-            clientGroupId: clientGroup.id,
-            pushVersion: 1,
-            schemaVersion: schema.version,
-            mutations: [mut],
-          });
-
-          expect(
-            await tx.query.replicacheClient.findFirst({
-              where: (t, { eq }) => eq(t.id, client.id),
-            }),
-          ).toMatchObject({ lastMutationId: mut.id });
-        },
-      );
-
-      await txTest(
-        `returns correct error for invalid schema version`,
-        async (tx) => {
-          const result = await push(tx, `1`, {
-            profileId: ``,
-            clientGroupId: ``,
-            pushVersion: 1,
-            schemaVersion: `666`,
-            mutations: [],
-          });
-
-          assert.deepEqual(result, {
-            error: `VersionNotSupported`,
-            versionType: `schema`,
-          });
-        },
-      );
-
-      await txTest(
-        `returns correct error for invalid push version`,
-        async (tx) => {
-          const result = await push(tx, `1`, {
-            profileId: ``,
-            clientGroupId: ``,
-            pushVersion: 666,
-            schemaVersion: schema.version,
-            mutations: [],
-          });
-
-          assert.deepEqual(result, {
-            error: `VersionNotSupported`,
-            versionType: `push`,
-          });
-        },
-      );
-    });
-  }
+    },
+  );
 });
 
 await test(`${pull.name} suite`, async (t) => {
-  for (const schema of supportedSchemas) {
-    await t.test(`schema ${schema.version}`, async (t) => {
-      const txTest = withTxTest(t, { isolationLevel: `repeatable read` });
-
-      await test(`database transaction isolation level`, async (t) => {
-        const txTest = withTxTest(t, { isolationLevel: `read committed` });
-
-        await txTest(`fails when using the default`, async (tx) => {
-          const result = pull(tx, `xxx`, {
-            profileId: ``,
-            clientGroupId: ``,
-            pullVersion: 1,
-            schemaVersion: schema.version,
-            cookie: null,
-          });
-
-          await assert.rejects(result, /transaction_isolation/);
-        });
-      });
-
-      await txTest(`creates a CVR with lastMutationIds`, async (tx) => {
-        const clientGroupId = nanoid();
-        const user = await createUser(tx);
-
-        await pull(tx, user.id, {
-          profileId: ``,
-          clientGroupId,
-          pullVersion: 1,
-          schemaVersion: schema.version,
-          cookie: null,
-        });
-
-        const clientGroup = await tx.query.replicacheClientGroup.findFirst({
-          where: (t, { eq }) => eq(t.userId, user.id),
-        });
-
-        expect(clientGroup).toMatchObject({ cvrVersion: 1 });
-      });
-
-      await txTest(
-        `non-existant client group creates one and stores cvrVersion`,
-        async (tx) => {
-          const user = await createUser(tx);
-
-          // Create a client group
-          const [clientGroup] = await tx
-            .insert(s.replicacheClientGroup)
-            .values([{ userId: user.id, schemaVersion: schema.version }])
-            .returning();
-          invariant(clientGroup != null);
-
-          // Create a client with a specific lastMutationId
-          const [client] = await tx
-            .insert(s.replicacheClient)
-            .values([{ clientGroupId: clientGroup.id, lastMutationId: 66 }])
-            .returning();
-          invariant(client != null);
-
-          const [skillState] = await tx
-            .insert(s.skillState)
-            .values([
-              {
-                userId: user.id,
-                srs: srsStateFromFsrsState(nextReview(null, Rating.Good)),
-                skill: glossToHanziWord(`我:i`),
-              },
-            ])
-            .returning();
-          invariant(skillState != null);
-
-          const result = await pull(tx, user.id, {
-            profileId: ``,
-            clientGroupId: clientGroup.id,
-            pullVersion: 1,
-            schemaVersion: schema.version,
-            cookie: null,
-          });
-
-          const cookie = `cookie` in result ? result.cookie : null;
-          assert.ok(cookie != null);
-
-          const cvr = await tx.query.replicacheCvr.findFirst({
-            where: (t, { eq }) => eq(t.id, cookie.cvrId),
-          });
-
-          const expectedEntities = await computeCvrEntities(
-            tx,
-            user.id,
-            schema,
-          );
-
-          // The CVR should have the lastMutationIds for the clients in the group
-          expect(cvr).toMatchObject({
-            lastMutationIds: { [client.id]: 66 },
-            entities: expectedEntities,
-          });
-        },
-      );
-
-      await txTest(
-        `returns lastMutationIdChanges only for changed clients`,
-        async (tx) => {
-          const clientGroupId = nanoid();
-          const clientId1 = nanoid();
-          const clientId2 = nanoid();
-
-          const user = await createUser(tx);
-
-          // Push a mutation from client 1
-          await push(tx, user.id, {
-            profileId: ``,
-            clientGroupId,
-            pushVersion: 1,
-            schemaVersion: schema.version,
-            mutations: [
-              {
-                id: 1,
-                name: `noop`,
-                args: {},
-                timestamp: 1,
-                clientId: clientId1,
-              },
-            ],
-          });
-
-          // A pull without a cookie should return all clients (at this point just
-          // client 1).
-          const pull1 = await pull(tx, user.id, {
-            profileId: ``,
-            clientGroupId,
-            pullVersion: 1,
-            schemaVersion: schema.version,
-            cookie: null,
-          });
-          assert.ok(`cookie` in pull1);
-          assert.deepEqual(pull1.lastMutationIdChanges, {
-            [clientId1]: 1,
-          });
-
-          // Do a new mutation from client 2
-          await push(tx, user.id, {
-            profileId: ``,
-            clientGroupId,
-            pushVersion: 1,
-            schemaVersion: schema.version,
-            mutations: [
-              {
-                id: 1,
-                name: `noop`,
-                args: {},
-                timestamp: 1,
-                clientId: clientId2,
-              },
-            ],
-          });
-
-          // A pull without a cookie should return all clients (now client 1 +
-          // client 2).
-          const pull2 = await pull(tx, user.id, {
-            profileId: ``,
-            clientGroupId,
-            pullVersion: 1,
-            schemaVersion: schema.version,
-            cookie: null,
-          });
-          assert.ok(`cookie` in pull2);
-          assert.deepEqual(pull2.lastMutationIdChanges, {
-            [clientId1]: 1,
-            [clientId2]: 1,
-          });
-
-          // A pull using cookie1 should only report client 2 as changed.
-          const pull3 = await pull(tx, user.id, {
-            profileId: ``,
-            clientGroupId,
-            pullVersion: 1,
-            schemaVersion: schema.version,
-            cookie: pull1.cookie,
-          });
-          assert.ok(`cookie` in pull3);
-          assert.deepEqual(pull3.lastMutationIdChanges, {
-            [clientId2]: 1,
-          });
-
-          // A pull using cookie3 should report no changes.
-          const pull4 = await pull(tx, user.id, {
-            profileId: ``,
-            clientGroupId,
-            pullVersion: 1,
-            schemaVersion: schema.version,
-            cookie: pull3.cookie,
-          });
-          assert.ok(`cookie` in pull4);
-          assert.deepEqual(pull4.lastMutationIdChanges, {});
-        },
-      );
-
-      await txTest(`null cookie, returns skillState patches`, async (tx) => {
-        const clientGroupId = nanoid();
-
-        const user = await createUser(tx);
-
-        const [skillState] = await tx
-          .insert(s.skillState)
-          .values([
-            {
-              userId: user.id,
-              srs: srsStateFromFsrsState(nextReview(null, Rating.Good)),
-              skill: glossToHanziWord(`我:i`),
-            },
-          ])
-          .returning();
-        invariant(skillState != null);
-
-        const result = await pull(tx, user.id, {
-          profileId: ``,
-          clientGroupId,
-          pullVersion: 1,
-          schemaVersion: schema.version,
-          cookie: null,
-        });
-
-        expect(result).toMatchObject({
-          cookie: {
-            order: 1,
-          },
-          patch: [
-            { op: `clear` },
-            {
-              op: `put`,
-              key: schema.skillState.marshalKey(skillState),
-              value: schema.skillState.marshalValue({
-                skill: skillState.skill,
-                srs: skillState.srs,
-              }),
-            },
-          ],
-        });
-      });
-
-      await txTest(`handles deletes for skillState`, async (tx) => {
-        const clientGroupId = nanoid();
-
-        const user = await createUser(tx);
-
-        const [skillState] = await tx
-          .insert(s.skillState)
-          .values([
-            {
-              userId: user.id,
-              srs: srsStateFromFsrsState(nextReview(null, Rating.Good)),
-              skill: glossToHanziWord(`我:i`),
-            },
-          ])
-          .returning();
-        invariant(skillState != null);
-
-        const pull1 = await pull(tx, user.id, {
-          profileId: ``,
-          clientGroupId,
-          pullVersion: 1,
-          schemaVersion: schema.version,
-          cookie: null,
-        });
-
-        await tx.delete(s.skillState).where(eq(s.skillState.id, skillState.id));
-
-        assert.ok(`cookie` in pull1);
-
-        const pull2 = await pull(tx, user.id, {
-          profileId: ``,
-          clientGroupId,
-          pullVersion: 1,
-          schemaVersion: schema.version,
-          cookie: pull1.cookie,
-        });
-
-        expect(pull2).toMatchObject({
-          cookie: {
-            order: 2,
-          },
-          patch: [
-            {
-              op: `del`,
-              key: schema.skillState.marshalKey(skillState),
-            },
-          ],
-        });
-      });
-
-      await txTest(`handles deletes for skillRating`, async (tx) => {
-        const clientGroupId = nanoid();
-
-        const user = await createUser(tx);
-
-        const now = new Date();
-
-        const [skillRating] = await tx
-          .insert(s.skillRating)
-          .values([
-            {
-              userId: user.id,
-              skill: glossToHanziWord(`我:i`),
-              rating: Rating.Good,
-              createdAt: now,
-            },
-          ])
-          .returning();
-        invariant(skillRating != null);
-
-        const pull1 = await pull(tx, user.id, {
-          profileId: ``,
-          clientGroupId,
-          pullVersion: 1,
-          schemaVersion: schema.version,
-          cookie: null,
-        });
-
-        await tx
-          .delete(s.skillRating)
-          .where(eq(s.skillRating.id, skillRating.id));
-
-        assert.ok(`cookie` in pull1);
-
-        const pull2 = await pull(tx, user.id, {
-          profileId: ``,
-          clientGroupId,
-          pullVersion: 1,
-          schemaVersion: schema.version,
-          cookie: pull1.cookie,
-        });
-
-        expect(pull2).toMatchObject({
-          cookie: {
-            order: 2,
-          },
-          patch: [
-            {
-              op: `del`,
-              key: schema.skillRating.marshalKey({ id: skillRating.id }),
-            },
-          ],
-        });
-      });
-
-      await txTest(`handles puts for skillRating`, async (tx) => {
-        const clientGroupId = nanoid();
-
-        const user = await createUser(tx);
-
-        const now = new Date();
-
-        const [skillRating] = await tx
-          .insert(s.skillRating)
-          .values([
-            {
-              userId: user.id,
-              skill: glossToHanziWord(`我:i`),
-              rating: Rating.Again,
-              createdAt: now,
-              durationMs: 1234,
-            },
-          ])
-          .returning();
-        invariant(skillRating != null);
-
-        const pull1 = await pull(tx, user.id, {
-          profileId: ``,
-          clientGroupId,
-          pullVersion: 1,
-          schemaVersion: schema.version,
-          cookie: null,
-        });
-
-        expect(pull1).toMatchObject({
-          patch: [
-            {
-              op: `clear`,
-            },
-            {
-              op: `put`,
-              key: schema.skillRating.marshalKey({ id: skillRating.id }),
-              value: schema.skillRating.marshalValue({
-                id: skillRating.id,
-                skill: glossToHanziWord(`我:i`),
-                rating: Rating.Again,
-                createdAt: now,
-                durationMs: 1234,
-              }),
-            },
-          ],
-        });
-      });
-    });
-  }
-});
-
-await test(`${withDbTest.name} examples`, async (t) => {
-  const dbTest = withDbTest(t);
-
-  await dbTest(`should work with inline fixtures 1`, async (db) => {
-    await db.insert(s.user).values([{ id: `1` }]);
-
-    // Your test logic here
-    const users = await db.select().from(s.user);
-    assert.equal(users.length, 1);
-  });
-});
-
-await test(`${computeCvrEntities.name} suite`, async (t) => {
-  for (const schema of supportedSchemas) {
-    await t.test(`schema ${schema.version}`, async (t) => {
-      const txTest = withTxTest(t);
-
-      await txTest(
-        `works for non-existant user and client group`,
-        async (tx) => {
-          assert.deepEqual(await computeCvrEntities(tx, `1`, schema), {
-            hanziGlossMistake: {},
-            hanziPinyinMistake: {},
-            pinyinFinalAssociation: {},
-            pinyinInitialAssociation: {},
-            pinyinInitialGroupTheme: {},
-            skillState: {},
-            skillRating: {},
-          });
-        },
-      );
-
-      await txTest(`works for user`, async (tx) => {
-        const user = await createUser(tx);
-
-        assert.deepEqual(await computeCvrEntities(tx, user.id, schema), {
-          hanziGlossMistake: {},
-          hanziPinyinMistake: {},
-          pinyinFinalAssociation: {},
-          pinyinInitialAssociation: {},
-          pinyinInitialGroupTheme: {},
-          skillState: {},
-          skillRating: {},
-        });
-      });
-
-      await txTest(`only includes skillState for the user`, async (tx) => {
-        const user1 = await createUser(tx);
-        const user2 = await createUser(tx);
-
-        const [user1SkillState] = await tx
-          .insert(s.skillState)
-          .values([
-            {
-              userId: user1.id,
-              skill: glossToHanziWord(`我:i`),
-              srs: srsStateFromFsrsState(nextReview(null, Rating.Good)),
-            },
-            {
-              userId: user2.id,
-              skill: glossToHanziWord(`我:i`),
-              srs: srsStateFromFsrsState(nextReview(null, Rating.Good)),
-            },
-          ])
-          .returning({
-            id: s.skillState.id,
-            skill: s.skillState.skill,
-            version: xmin(s.skillState),
-          });
-        invariant(user1SkillState != null);
-
-        assert.deepEqual(await computeCvrEntities(tx, user1.id, schema), {
-          hanziGlossMistake: {},
-          hanziPinyinMistake: {},
-          pinyinFinalAssociation: {},
-          pinyinInitialAssociation: {},
-          pinyinInitialGroupTheme: {},
-          skillRating: {},
-          skillState: {
-            [user1SkillState.id]:
-              user1SkillState.version +
-              `:` +
-              schema.skillState.marshalKey(user1SkillState),
-          },
-        });
-      });
-
-      await txTest(`only includes skillRating for the user`, async (tx) => {
-        const user1 = await createUser(tx);
-        const user2 = await createUser(tx);
-
-        const [user1SkillRating] = await tx
-          .insert(s.skillRating)
-          .values([
-            {
-              userId: user1.id,
-              skill: glossToHanziWord(`我:i`),
-              rating: Rating.Again,
-            },
-            {
-              userId: user2.id,
-              skill: glossToHanziWord(`我:i`),
-              rating: Rating.Good,
-            },
-          ])
-          .returning({
-            id: s.skillRating.id,
-            skill: s.skillRating.skill,
-            createdAt: s.skillRating.createdAt,
-            version: xmin(s.skillRating),
-          });
-        invariant(user1SkillRating != null);
-
-        assert.deepEqual(await computeCvrEntities(tx, user1.id, schema), {
-          hanziGlossMistake: {},
-          hanziPinyinMistake: {},
-          pinyinFinalAssociation: {},
-          pinyinInitialAssociation: {},
-          pinyinInitialGroupTheme: {},
-          skillRating: {
-            [user1SkillRating.id]:
-              user1SkillRating.version +
-              `:` +
-              schema.skillRating.marshalKey({ id: user1SkillRating.id }),
-          },
-          skillState: {},
-        });
-      });
-
-      await txTest(
-        `only includes pinyinFinalAssociation for the user`,
-        async (tx) => {
-          const user1 = await createUser(tx);
-          const user2 = await createUser(tx);
-
-          const [user1PinyinFinalAssociation] = await tx
-            .insert(s.pinyinFinalAssociation)
-            .values([
-              {
-                userId: user1.id,
-                final: `p`,
-                name: `p1`,
-              },
-              {
-                userId: user2.id,
-                final: `p`,
-                name: `p2`,
-              },
-            ])
-            .returning({
-              id: s.pinyinFinalAssociation.id,
-              final: s.pinyinFinalAssociation.final,
-              version: xmin(s.pinyinFinalAssociation),
-            });
-          invariant(user1PinyinFinalAssociation != null);
-
-          assert.deepEqual(await computeCvrEntities(tx, user1.id, schema), {
-            hanziGlossMistake: {},
-            hanziPinyinMistake: {},
-            pinyinFinalAssociation: {
-              [user1PinyinFinalAssociation.id]:
-                user1PinyinFinalAssociation.version +
-                `:` +
-                schema.pinyinFinalAssociation.marshalKey(
-                  user1PinyinFinalAssociation,
-                ),
-            },
-            pinyinInitialAssociation: {},
-            pinyinInitialGroupTheme: {},
-            skillRating: {},
-            skillState: {},
-          });
-        },
-      );
-
-      await txTest(
-        `only includes pinyinInitialAssociation for the user`,
-        async (tx) => {
-          const user1 = await createUser(tx);
-          const user2 = await createUser(tx);
-
-          const [user1PinyinInitialAssociation] = await tx
-            .insert(s.pinyinInitialAssociation)
-            .values([
-              {
-                userId: user1.id,
-                initial: `p`,
-                name: `p1`,
-              },
-              {
-                userId: user2.id,
-                initial: `p`,
-                name: `p2`,
-              },
-            ])
-            .returning({
-              id: s.pinyinInitialAssociation.id,
-              initial: s.pinyinInitialAssociation.initial,
-              version: xmin(s.pinyinInitialAssociation),
-            });
-          invariant(user1PinyinInitialAssociation != null);
-
-          assert.deepEqual(await computeCvrEntities(tx, user1.id, schema), {
-            hanziGlossMistake: {},
-            hanziPinyinMistake: {},
-            pinyinFinalAssociation: {},
-            pinyinInitialAssociation: {
-              [user1PinyinInitialAssociation.id]:
-                user1PinyinInitialAssociation.version +
-                `:` +
-                schema.pinyinInitialAssociation.marshalKey(
-                  user1PinyinInitialAssociation,
-                ),
-            },
-            pinyinInitialGroupTheme: {},
-            skillRating: {},
-            skillState: {},
-          });
-        },
-      );
-    });
-  }
-});
-
-await test(`${fetchMutations.name} suite`, async (t) => {
   const txTest = withTxTest(t, { isolationLevel: `repeatable read` });
 
-  await txTest(`works for non-existant user and client group`, async (tx) => {
-    const clientGroupId = nanoid();
-    const clientId = nanoid();
+  await txTest(
+    `returns correct error for invalid schema version`,
+    async (tx) => {
+      const result = await pull(tx, `5`, {
+        profileId: ``,
+        clientGroupId: ``,
+        pullVersion: 1,
+        schemaVersion: `666`,
+        cookie: null,
+      });
 
-    const user = await createUser(tx);
-
-    // Push a mutation from client 1
-    await push(tx, user.id, {
-      profileId: ``,
-      clientGroupId,
-      pushVersion: 1,
-      schemaVersion: v7.version,
-      mutations: [
-        {
-          id: 1,
-          name: `noop`,
-          args: { arg1: `value1` },
-          timestamp: 1,
-          clientId,
-        },
-      ],
-    });
-
-    assert.deepEqual(
-      await fetchMutations(tx, user.id, {
-        schemaVersions: [v7.version],
-        lastMutationIds: {},
-      }),
-      {
-        mutations: [
-          {
-            clientGroupId,
-            mutations: [
-              {
-                args: { arg1: `value1` },
-                clientId,
-                id: 1,
-                name: `noop`,
-                timestamp: 1,
-              },
-            ],
-            schemaVersion: v7.version,
-          },
-        ],
-      },
-    );
-  });
+      assert.deepEqual(result, {
+        error: `VersionNotSupported`,
+        versionType: `schema`,
+      });
+    },
+  );
 });
